Fully working logic.
The cleaning logic fully works with this commit. The cleaned dataset (available as a dictionary) is not easily accessible yet. I accessed the private class variables to complete my minor homework assignment. Some of the logic is messy such as the strings and queries spilled all over the code. I intend to separate these out into a data class or something later.